Solve one circular dep of generator
