+++ packages/generator/src/edmx-parser/v4/edmx-parser.ts
@@ -6,7 +6,6 @@ import {
   parseEntityTypesBase
 } from '../common/edmx-parser';
 import { forceArray } from '../../generator-utils';
-import { joinEntityTypes } from '../../edmx-to-vdm/v4/entity';
 import { stripNamespace } from '../../edmx-to-vdm/edmx-to-vdm-util';
 import {
   EdmxAction,
@@ -21,6 +20,26 @@ import {
   EdmxNavigationPropertyBinding
 } from './edm-types';
 
+/**
+ * @internal
+ */
+ export function joinEntityTypes<T extends EdmxEntityTypeV4>(
+  entityType: T,
+  baseType: T
+): T {
+  // TODO: only join properties / nav properties of the respective type
+  return {
+    ...entityType,
+    Key: {
+      PropertyRef: [...entityType.Key.PropertyRef, ...baseType.Key.PropertyRef]
+    },
+    Property: [...entityType.Property, ...baseType.Property],
+    NavigationProperty: [
+      ...entityType.NavigationProperty,
+      ...baseType.NavigationProperty
+    ]
+  };
+}
 /**
  * @internal
  */
